refactor: 把 1798 行 src/index.js 拆成多個 CommonJS 模組
src/index.js 從 1798 行縮到 143 行，只保留 client init + event handler
dispatcher。其他職責拆到 17 個 module，每個檔案單一職責。

模組結構：
- config.js — 所有 env vars 與常數
- url-routing.js — host sets / normalizeUrl / extractSupportedUrls / isXxxUrl / fixers
- utils.js — trimDescription / pickRandom
- probe.js — playwright subprocess wrapper + threads metadata cache
- embeds.js — 所有 EmbedBuilder factory
- platforms/{threads,instagram,bilibili,bahamut,ptt}.js — 各平台 payload builder
- preview.js — buildPreviewPayloads 平台分派 dispatcher
- discord-io.js — send/suppress/empty-embed/dedup state/permissions
- ai/{persona,memory,providers,chain}.js — 西寶 AI 子系統
- mention.js — @西寶 dispatcher（抽籤/道歉/AI/hardcoded fallback）
- commands.js — slash commands

驗證手段：
- scripts/smoke.js（新增）— 純函式 assertion 測試 44 條，含 normalizeUrl
  各 host 的 tracking-param 行為、buildFallbackUrl 各平台路由、
  buildUserTurn 名稱優先順序、Gemini/OpenAI message 格式轉換等
- node --check 通過 src/ + scripts/ 全部 20 個檔
- 每個 module 獨立 require() 載入無錯
- index.js 改用 require.main === module 包住 client.login，方便測試 require

CLAUDE.md 同步更新 architecture 區塊以新結構為準。

行為保留：所有 platform routing、AI provider chain 順序、persona、
empty-embed handling、dedup window、threads multi-image 在 video 前判斷的
PR #12 行為，全數保留。
