+++ src/ai/providers.js
@@ -1,0 +1,225 @@
+const {
+  AI_TIMEOUT_MS,
+  AI_PERSONA,
+  GEMINI_API_KEY,
+  GEMINI_MODEL,
+  GROQ_API_KEY,
+  CEREBRAS_API_KEY,
+  CEREBRAS_MODEL,
+  DEEPSEEK_API_KEY,
+  DEEPSEEK_MODEL,
+} = require("../config");
+const { buildOpenAIMessages, buildGeminiContents } = require("./persona");
+
+async function withAbortTimeout(timeoutMs, providerLabel, fn) {
+  const controller = new AbortController();
+  const timer = setTimeout(() => controller.abort(), timeoutMs);
+  try {
+    return await fn(controller.signal);
+  } catch (error) {
+    if (error.name === "AbortError") {
+      console.warn(`[ai] ${providerLabel} timed out after ${timeoutMs}ms`);
+    } else {
+      console.warn(`[ai] ${providerLabel} failed: ${error.message}`);
+    }
+    return null;
+  } finally {
+    clearTimeout(timer);
+  }
+}
+
+function logRateHeaders(label, response) {
+  const remainingTokens = response.headers.get("x-ratelimit-remaining-tokens");
+  const remainingRequests = response.headers.get("x-ratelimit-remaining-requests");
+  if (remainingTokens || remainingRequests) {
+    console.log(
+      `[ai] ${label} remaining tokens=${remainingTokens ?? "?"} req=${remainingRequests ?? "?"}`,
+    );
+  }
+}
+
+async function callGemini(turns) {
+  const endpoint = `https://generativelanguage.googleapis.com/v1beta/models/${encodeURIComponent(
+    GEMINI_MODEL,
+  )}:generateContent?key=${encodeURIComponent(GEMINI_API_KEY)}`;
+
+  const body = {
+    system_instruction: { parts: [{ text: AI_PERSONA }] },
+    contents: buildGeminiContents(turns),
+    generationConfig: { temperature: 0.9, topP: 0.95, maxOutputTokens: 180 },
+  };
+
+  return withAbortTimeout(AI_TIMEOUT_MS, "gemini", async (signal) => {
+    const response = await fetch(endpoint, {
+      method: "POST",
+      headers: { "Content-Type": "application/json" },
+      body: JSON.stringify(body),
+      signal,
+    });
+
+    if (!response.ok) {
+      const errText = await response.text().catch(() => "");
+      console.warn(`[ai] gemini http ${response.status}: ${errText.slice(0, 200)}`);
+      return null;
+    }
+
+    const payload = await response.json();
+    const text = payload?.candidates?.[0]?.content?.parts
+      ?.map((p) => p.text || "")
+      .join("")
+      .trim();
+
+    if (!text) {
+      const finishReason = payload?.candidates?.[0]?.finishReason ?? "unknown";
+      console.warn(`[ai] gemini empty response, finishReason=${finishReason}`);
+      return null;
+    }
+    return text;
+  });
+}
+
+async function callGroq(turns, model) {
+  const body = {
+    model,
+    messages: buildOpenAIMessages(turns),
+    temperature: 0.9,
+    top_p: 0.95,
+    max_tokens: 180,
+  };
+  const label = `groq:${model}`;
+
+  return withAbortTimeout(AI_TIMEOUT_MS, label, async (signal) => {
+    const response = await fetch("https://api.groq.com/openai/v1/chat/completions", {
+      method: "POST",
+      headers: {
+        "Content-Type": "application/json",
+        Authorization: `Bearer ${GROQ_API_KEY}`,
+      },
+      body: JSON.stringify(body),
+      signal,
+    });
+
+    logRateHeaders(label, response);
+
+    if (!response.ok) {
+      const errText = await response.text().catch(() => "");
+      console.warn(`[ai] ${label} http ${response.status}: ${errText.slice(0, 200)}`);
+      return null;
+    }
+
+    const payload = await response.json();
+    const text = payload?.choices?.[0]?.message?.content?.trim();
+    if (!text) {
+      const finishReason = payload?.choices?.[0]?.finish_reason ?? "unknown";
+      console.warn(`[ai] ${label} empty response, finishReason=${finishReason}`);
+      return null;
+    }
+    return text;
+  });
+}
+
+const CEREBRAS_QUEUE_RETRY_DELAY_MS = 1500;
+async function callCerebras(turns) {
+  const body = {
+    model: CEREBRAS_MODEL,
+    messages: buildOpenAIMessages(turns),
+    temperature: 0.9,
+    top_p: 0.95,
+    max_tokens: 180,
+  };
+  const label = `cerebras:${CEREBRAS_MODEL}`;
+
+  return withAbortTimeout(AI_TIMEOUT_MS, label, async (signal) => {
+    for (let attempt = 0; attempt < 2; attempt++) {
+      const response = await fetch("https://api.cerebras.ai/v1/chat/completions", {
+        method: "POST",
+        headers: {
+          "Content-Type": "application/json",
+          Authorization: `Bearer ${CEREBRAS_API_KEY}`,
+        },
+        body: JSON.stringify(body),
+        signal,
+      });
+
+      logRateHeaders(label, response);
+
+      if (response.ok) {
+        const payload = await response.json();
+        const text = payload?.choices?.[0]?.message?.content?.trim();
+        if (!text) {
+          const finishReason = payload?.choices?.[0]?.finish_reason ?? "unknown";
+          console.warn(`[ai] ${label} empty response, finishReason=${finishReason}`);
+          return null;
+        }
+        if (attempt > 0) console.log(`[ai] ${label} succeeded on retry`);
+        return text;
+      }
+
+      const errText = await response.text().catch(() => "");
+      const isQueueExceeded =
+        response.status === 429 &&
+        (errText.includes("queue_exceeded") || errText.includes("high traffic"));
+
+      if (isQueueExceeded && attempt === 0) {
+        console.log(
+          `[ai] ${label} queue_exceeded, retrying in ${CEREBRAS_QUEUE_RETRY_DELAY_MS}ms`,
+        );
+        await new Promise((resolve) => setTimeout(resolve, CEREBRAS_QUEUE_RETRY_DELAY_MS));
+        continue;
+      }
+
+      console.warn(`[ai] ${label} http ${response.status}: ${errText.slice(0, 200)}`);
+      return null;
+    }
+    return null;
+  });
+}
+
+async function callDeepSeek(turns) {
+  const body = {
+    model: DEEPSEEK_MODEL,
+    messages: buildOpenAIMessages(turns),
+    temperature: 0.9,
+    top_p: 0.95,
+    max_tokens: 180,
+  };
+  const label = `deepseek:${DEEPSEEK_MODEL}`;
+
+  return withAbortTimeout(AI_TIMEOUT_MS, label, async (signal) => {
+    const response = await fetch("https://api.deepseek.com/chat/completions", {
+      method: "POST",
+      headers: {
+        "Content-Type": "application/json",
+        Authorization: `Bearer ${DEEPSEEK_API_KEY}`,
+      },
+      body: JSON.stringify(body),
+      signal,
+    });
+
+    logRateHeaders(label, response);
+
+    if (!response.ok) {
+      const errText = await response.text().catch(() => "");
+      console.warn(`[ai] ${label} http ${response.status}: ${errText.slice(0, 200)}`);
+      return null;
+    }
+
+    const payload = await response.json();
+    const text = payload?.choices?.[0]?.message?.content?.trim();
+    if (!text) {
+      const finishReason = payload?.choices?.[0]?.finish_reason ?? "unknown";
+      console.warn(`[ai] ${label} empty response, finishReason=${finishReason}`);
+      return null;
+    }
+    return text;
+  });
+}
+
+module.exports = {
+  withAbortTimeout,
+  logRateHeaders,
+  callGemini,
+  callGroq,
+  callCerebras,
+  callDeepSeek,
+};
