+++ src/utils.js
@@ -1,0 +1,12 @@
+function trimDescription(text, limit) {
+  if (!text || text.length <= limit) {
+    return text;
+  }
+  return `${text.slice(0, limit - 1).trimEnd()}…`;
+}
+
+function pickRandom(arr) {
+  return arr[Math.floor(Math.random() * arr.length)];
+}
+
+module.exports = { trimDescription, pickRandom };
